# api-spec.yaml
openapi: 3.0.0
info:
  title: Groov API
  version: 1.0.0
  description: |
    REST API specifikacija Groov šokių judesių mokymosi ir choreografijos platformai.
    Groov sistema skirta šokėjams padėti mokytis ir atsiminti šokių judesius,
    kurti sekų choreografijas ir įvertinti kitų vartotojų sukurtas sekas.

    **Autentifikacija:**
    - Sistema naudoja JWT token'us su HTTP-only cookies
    - Access token'ai ir refresh token'ai saugomi slapukuose
    - Chranomi endpointai pažymėti 🔒 simboliu

  contact:
    name: Groov Support
    email: support@groov.lt
servers:
  - url: http://localhost:3003
    description: Local development server
  - url: https://api.groov.lt
    description: Production server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT access token saugomas HTTP-only cookie
  schemas:
...
    DanceMove:
      type: object
      required:
        - name
        - difficulty
        - start_position
        - end_position
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Swing Out"
          description: Šokio judesiuo pavadinimas
        description:
          type: string
          maxLength: 500
          example: "Basic swing out movement from closed position to open"
          description: Detalus judesio aprašymas
        difficulty:
          type: string
          enum: [easy, medium, hard, very_hard]
          example: "medium"
          description: Judesio sudėtingumas
        start_position:
          type: string
          enum: [closed, openLeftToRight, openRightToRight, openLeftToLeft, openRightToLeft, sweethearts]
          example: "closed"
          description: Pradinė pozicija prieš judesį
        end_position:
          type: string
          enum: [closed, openLeftToRight, openRightToRight, openLeftToLeft, openRightToLeft, sweethearts]
          example: "openLeftToRight"
          description: Galinė pozicija po judesio
        parent_move_id:
          type: integer
          nullable: true
          example: null
          description: Tėvinio judesio ID (jei judesis yra variacija)
        created_at:
          type: string
          format: date-time
          readOnly: true
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          readOnly: true
          example: "2024-01-15T10:30:00Z"
        deleted:
          type: boolean
          readOnly: true
          example: false
    DanceMoveCreate:
      type: object
      required:
        - name
        - difficulty
        - start_position
        - end_position
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Swing Out"
        description:
          type: string
          maxLength: 500
          example: "Basic swing out movement from closed position to open"
        difficulty:
          type: string
          enum: [easy, medium, hard, very_hard]
          example: "medium"
        start_position:
          type: string
          enum: [closed, openLeftToRight, openRightToRight, openLeftToLeft, openRightToLeft, sweethearts]
          example: "closed"
        end_position:
          type: string
          enum: [closed, openLeftToRight, openRightToRight, openLeftToLeft, openRightToLeft, sweethearts]
          example: "openLeftToRight"
        parent_move_id:
          type: integer
          nullable: true
          example: null
    DanceMoveUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Swing Out Updated"
        description:
          type: string
          maxLength: 500
          example: "Updated description"
        difficulty:
          type: string
          enum: [easy, medium, hard, very_hard]
          example: "hard"
        start_position:
          type: string
          enum: [closed, openLeftToRight, openRightToRight, openLeftToLeft, openRightToLeft, sweethearts]
          example: "closed"
        end_position:
          type: string
          enum: [closed, openLeftToRight, openRightToRight, openLeftToLeft, openRightToLeft, sweethearts]
          example: "openLeftToRight"
        parent_move_id:
          type: integer
          nullable: true
          example: 5

    # Dance Sequence Schemas
...
paths:
...
  /dance-moves:
    get:
      tags:
        - Dance Moves
      summary: Gauti visus šokių judesius
      description: |
        Grąžina visų aktyvių (neištrintų) šokių judesių sąrašą.
        Galima filtruoti pagal pavadinimą naudojant search parametrą.
      operationId: getAllDanceMoves
      parameters:
        - name: search
          in: query
          description: Paieškos tekstas judesio pavadinime
          required: false
          schema:
            type: string
          example: "swing"
      responses:
        '200':
          description: Šokių judesių sąrašas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DanceMove'
              examples:
                movesList:
                  summary: Judesių sąrašo pavyzdys
                  value:
                    - id: 1
                      name: "Swing Out"
                      description: "Basic swing out movement"
                      difficulty: "easy"
                      start_position: "closed"
                      end_position: "openLeftToRight"
                      parent_move_id: null
                      created_at: "2024-01-15T10:30:00Z"
                      updated_at: "2024-01-15T10:30:00Z"
                      deleted: false
                    - id: 2
                      name: "Tuck Turn"
                      description: "Turn variation from open position"
                      difficulty: "medium"
                      start_position: "openLeftToRight"
                      end_position: "openLeftToRight"
                      parent_move_id: null
                      created_at: "2024-01-15T11:00:00Z"
                      updated_at: "2024-01-15T11:00:00Z"
                      deleted: false
      x-codeSamples:
        - lang: 'curl'
          source: |
            # Gauti visus judesius
            curl -X GET http://localhost:3003/dance-moves

            # Su paieška
            curl -X GET "http://localhost:3003/dance-moves?search=swing"
        - lang: 'JavaScript'
          source: |
            // Gauti visus judesius
            const response = await fetch('http://localhost:3003/dance-moves');
            const moves = await response.json();

            // Su paieška
            const searchResponse = await fetch('http://localhost:3003/dance-moves?search=swing');
            const searchResults = await searchResponse.json();

    post:
      tags:
        - Dance Moves
      summary: 🔒 Sukurti naują šokių judesį
      description: |
        Sukuria naują šokių judesį sistemoje.
        **Reikalinga admin teisės.**
      operationId: createDanceMove
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DanceMoveCreate'
            examples:
              basicMove:
                summary: Paprasto judesio kūrimas
                value:
                  name: "Circle Left"
                  description: "Circular movement to the left"
                  difficulty: "easy"
                  start_position: "closed"
                  end_position: "closed"
                  parent_move_id: null
              variationMove:
                summary: Variacija su tėviniu judesiu
                value:
                  name: "Swing Out with Style"
                  description: "Stylish variation of basic swing out"
                  difficulty: "hard"
                  start_position: "closed"
                  end_position: "openLeftToRight"
                  parent_move_id: 1
      responses:
        '201':
          description: Judesis sukurtas sėkmingai
          headers:
            Location:
              schema:
                type: string
              description: URL naujai sukurto judesio
              example: "/dance-moves/15"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DanceMove'
              example:
                id: 15
                name: "Circle Left"
                description: "Circular movement to the left"
                difficulty: "easy"
                start_position: "closed"
                end_position: "closed"
                parent_move_id: null
                created_at: "2024-01-20T14:30:00Z"
                updated_at: "2024-01-20T14:30:00Z"
                deleted: false
        '400':
          description: Neteisingi duomenys arba validacijos klaida
          content:
            application/json:
              examples:
                emptyBody:
                  summary: Tuščias request body
                  value:
                    error: "Bad request"
                    message: "Request body cannot be empty."
                idNotAllowed:
                  summary: ID laukas neleistinas
                  value:
                    error: "Bad request"
                    message: "ID field is not allowed when creating a new DanceMove."
                validationError:
                  summary: Validacijos klaida
                  value:
                    error: "Validation error"
                    message: "Could not post dance move object"
        '401':
          description: Neautentifikuotas vartotojas
          content:
            application/json:
              example:
                error: "Unauthorized"
                message: "Authentication required"
        '403':
          description: Nepakankamos teisės (ne admin)
          content:
            application/json:
              example:
                error: "Forbidden"
                message: "Admin role required"
      x-codeSamples:
        - lang: 'curl'
          source: |
            curl -X POST http://localhost:3003/dance-moves \
              -H "Content-Type: application/json" \
              -b cookies.txt \
              -d '{
                "name": "Circle Left",
                "description": "Circular movement to the left",
                "difficulty": "easy",
                "start_position": "closed",
                "end_position": "closed",
                "parent_move_id": null
              }'
        - lang: 'JavaScript'
          source: |
            const response = await fetch('http://localhost:3003/dance-moves', {
              method: 'POST',
              headers: { 'Content-Type': 'application/json' },
              credentials: 'include',
              body: JSON.stringify({
                name: 'Circle Left',
                description: 'Circular movement to the left',
                difficulty: 'easy',
                start_position: 'closed',
                end_position: 'closed',
                parent_move_id: null
              })
            });
            const newMove = await response.json();
  /dance-moves/{id}:
    get:
      tags:
        - Dance Moves
      summary: Gauti šokių judesį pagal ID
      description: Grąžina detalią informaciją apie konkretų šokių judesį
      operationId: getDanceMoveById
      parameters:
        - name: id
          in: path
          required: true
          description: Šokių judesio ID
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: Šokių judesio informacija
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DanceMove'
              example:
                id: 1
                name: "Swing Out"
                description: "Basic swing out movement from closed to open position"
                difficulty: "easy"
                start_position: "closed"
                end_position: "openLeftToRight"
                parent_move_id: null
                created_at: "2024-01-15T10:30:00Z"
                updated_at: "2024-01-15T10:30:00Z"
                deleted: false
        '400':
          description: Neteisingas ID formatas
          content:
            application/json:
              example:
                error: "Bad request"
                message: "DanceMove id has to be a positive integer"
        '404':
          description: Judesis nerastas
          content:
            application/json:
              example:
                error: "Not found"
                message: "Dance move not found"
      x-codeSamples:
        - lang: 'curl'
          source: |
            curl -X GET http://localhost:3003/dance-moves/1
        - lang: 'JavaScript'
          source: |
            const response = await fetch('http://localhost:3003/dance-moves/1');
            const move = await response.json();

    put:
      tags:
        - Dance Moves
      summary: 🔒 Atnaujinti šokių judesį
      description: |
        Atnaujina egzistuojančio šokių judesio informaciją.
        **Reikalinga admin teisės.**
      operationId: updateDanceMove
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Šokių judesio ID
          schema:
            type: integer
            minimum: 1
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DanceMoveUpdate'
            example:
              name: "Swing Out - Updated"
              description: "Updated description with more details"
              difficulty: "medium"
      responses:
        '200':
          description: Judesis atnaujintas sėkmingai
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DanceMove'
              example:
                id: 1
                name: "Swing Out - Updated"
                description: "Updated description with more details"
                difficulty: "medium"
                start_position: "closed"
                end_position: "openLeftToRight"
                parent_move_id: null
                created_at: "2024-01-15T10:30:00Z"
                updated_at: "2024-01-20T15:00:00Z"
                deleted: false
        '400':
          description: Neteisingi duomenys
          content:
            application/json:
              examples:
                invalidId:
                  value:
                    error: "Bad request"
                    message: "DanceMove id has to be a positive integer"
                idInBody:
                  value:
                    error: "Bad request"
                    message: "ID field is not allowed when updating a new DanceMove."
        '401':
          description: Neautentifikuotas
          content:
            application/json:
              example:
                error: "Unauthorized"
        '403':
          description: Nepakankamos teisės
          content:
            application/json:
              example:
                error: "Forbidden"
                message: "Admin role required"
        '404':
          description: Judesis nerastas
          content:
            application/json:
              example:
                error: "Not found"
      x-codeSamples:
        - lang: 'curl'
          source: |
            curl -X PUT http://localhost:3003/dance-moves/1 \
              -H "Content-Type: application/json" \
              -b cookies.txt \
              -d '{
                "name": "Swing Out - Updated",
                "difficulty": "medium"
              }'
        - lang: 'JavaScript'
          source: |
            const response = await fetch('http://localhost:3003/dance-moves/1', {
              method: 'PUT',
              headers: { 'Content-Type': 'application/json' },
              credentials: 'include',
              body: JSON.stringify({
                name: 'Swing Out - Updated',
                difficulty: 'medium'
              })
            });

    delete:
      tags:
        - Dance Moves
      summary: 🔒 Ištrinti šokių judesį
      description: |
        Atliekamas soft delete - pažymi judesį kaip ištrintą, bet nepašalina iš duomenų bazės.
        **Reikalinga admin teisės.**
      operationId: deleteDanceMove
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Šokių judesio ID
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: Judesis ištrintas sėkmingai
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DanceMove'
              example:
                id: 1
                name: "Swing Out"
                deleted: true
        '400':
          description: Neteisingas ID
          content:
            application/json:
              example:
                error: "Bad request"
                message: "DanceMove id has to be a positive integer"
        '401':
          description: Neautentifikuotas
        '403':
          description: Nepakankamos teisės
        '404':
          description: Judesis nerastas
          content:
            application/json:
              example:
                error: "Not found"
      x-codeSamples:
        - lang: 'curl'
          source: |
            curl -X DELETE http://localhost:3003/dance-moves/1 \
              -b cookies.txt
        - lang: 'JavaScript'
          source: |
            const response = await fetch('http://localhost:3003/dance-moves/1', {
              method: 'DELETE',
              credentials: 'include'
            });
  /dance-moves/difficulty/{difficulty}:
    get:
      tags:
        - Dance Moves
      summary: Gauti judesius pagal sudėtingumą
      description: Filtruoja šokių judesius pagal sudėtingumo lygį
      operationId: getDanceMovesByDifficulty
      parameters:
        - name: difficulty
          in: path
          required: true
          description: Sudėtingumo lygis
          schema:
            type: string
            enum: [easy, medium, hard, very_hard]
          example: "easy"
      responses:
        '200':
          description: Filtruotų judesių sąrašas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DanceMove'
              example:
                - id: 1
                  name: "Swing Out"
                  difficulty: "easy"
                  start_position: "closed"
                  end_position: "openLeftToRight"
                - id: 5
                  name: "Circle Left"
                  difficulty: "easy"
                  start_position: "closed"
                  end_position: "closed"
        '400':
          description: Neteisingas sudėtingumo lygis
          content:
            application/json:
              example:
                error: "Bad request"
                message: "Invalid difficulty value: ultra_hard"
      x-codeSamples:
        - lang: 'curl'
          source: |
            curl -X GET http://localhost:3003/dance-moves/difficulty/easy
        - lang: 'JavaScript'
          source: |
            const response = await fetch('http://localhost:3003/dance-moves/difficulty/easy');
            const easyMoves = await response.json();
  /dance-moves/{id}/children:
    get:
      tags:
        - Dance Moves
      summary: Gauti judesio variacijų sąrašą
      description: Grąžina visus judesius, kurie yra nurodyto judesio variacijomis (children)
      operationId: getChildMoves
      parameters:
        - name: id
          in: path
          required: true
          description: Tėvinio judesio ID
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: Vaikiškų judesių sąrašas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DanceMove'
              example:
                - id: 10
                  name: "Swing Out with Styling"
                  parent_move_id: 1
                  difficulty: "hard"
                - id: 11
                  name: "Swing Out Delayed"
                  parent_move_id: 1
                  difficulty: "medium"
        '400':
          description: Neteisingas ID
        '404':
          description: Tėvinis judesis nerastas
      x-codeSamples:
        - lang: 'curl'
          source: |
            curl -X GET http://localhost:3003/dance-moves/1/children
        - lang: 'JavaScript'
          source: |
            const response = await fetch('http://localhost:3003/dance-moves/1/children');
            const variations = await response.json();
  /dance-moves/{id}/parent:
    get:
      tags:
        - Dance Moves
      summary: Gauti tėvinį judesį
      description: Grąžina tėvinį judesį, jei judesis yra variacija
      operationId: getParentMove
      parameters:
        - name: id
          in: path
          required: true
          description: Vaiko judesio ID
          schema:
            type: integer
            minimum: 1
          example: 10
      responses:
        '200':
          description: Tėvinis judesis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DanceMove'
              example:
                id: 1
                name: "Swing Out"
                parent_move_id: null
        '400':
          description: Neteisingas ID
        '404':
          description: Judesis neturi tėvinio judesio arba nerastas
      x-codeSamples:
        - lang: 'curl'
          source: |
            curl -X GET http://localhost:3003/dance-moves/10/parent
        - lang: 'JavaScript'
          source: |
            const response = await fetch('http://localhost:3003/dance-moves/10/parent');
            const parent = await response.json();

  # ============================================
  # DANCE SEQUENCES ENDPOINTS
  # ============================================
...